fix: shared-android.gradle.kts red in AS Chipmunk
When updating to Android Studio Chimpmunk, after removing all build directories, the `android` section of `shared-android.gradle.kts` was unrecoginized.

I found and applied this fix:
https://github.com/Gurupreet/ComposeCookBook/issues/112#issuecomment-988882833

In the process of testing this, I discovered that AS was launching both Intel and M1 java processes. Turned out that the 1.8 java processes were Intel, so I switched all `kotlin.jvmTargets` to 11 to avoid this.