fix: long-running effects cause hangs
The previous implementation of `sink` used `toList` to accumulate the list of effects to proccess, but this requires that the flow completes before any actions are sent back to the store. Long-running effects - especially those that never complete - will cause the store to hang waiting for the flow to complete.

The solution is to replace the no-arg `sink()` with one that takes a `FlowCollector` so that effects can be processed as they come in. This required similar changes to Effect cancellation and `TestStore`.

`TestStore` was also modified to make its `assert` block a `suspend` block so that suspend functions can be called from tests.

Also added a convenience extension function `Flow.asEffect()` as a analog to Swift TCA's `Publisher.eraseToEffect()`