chore: switch back to Java 8
Much about this is unclear to me, but when testing on Android Studio Dolphin and Electric Eel, the problem of the Android section of build.gradle turning red reoccured. Previously, I had worked around this by switching to Java 11. This time, I was able to fix it by switching back to Java 8.

Possibly related, I was having issues with the Preferences menu being missing after updating to AS Chipmunk. Google searching suggested it was due to AS using an incorrect JRE version. Jetbrains recommends not changing JRE from the embedded version. I don't recall changing this, and when opening "Choose Boot Runtime for the IDE", the embedded version was set. But after "Use Default Runtime" and restarting, the prefs menu came back. Maybe this is related to the above issue somehow?